+++ R/inferrnal.R
@@ -27,6 +27,7 @@ parse_stockholm_msa_chunk <- function(x, pos, acc) {
 #' lines.  Other annotations are ignored.
 #'
 #' @param stockholm (\code{character} scalar) Path to a file to parse
+#' @param dna (\code{logical} scalar) Parse the input as DNA instead of RNA.
 #'
 #' @return a \code{list}.  The sequences themselves are element
 #'     "\code{alignment}", and \code{#=GC} annotations are given as
@@ -42,10 +43,11 @@ parse_stockholm_msa_chunk <- function(x, pos, acc) {
 #'     msa$SS_cons
 #'     # reference sequence
 #'     msa$RF
-read_stockholm_msa <- function(stockholm) {
+read_stockholm_msa <- function(stockholm, dna = FALSE) {
     assertthat::assert_that((assertthat::is.string(stockholm) &&
                                 file.exists(stockholm)) ||
-                                methods::is(stockholm, "connection"))
+                                methods::is(stockholm, "connection"),
+                            assertthat::is.flag(dna))
 
     seqs <-
         readr::read_lines_chunked(
@@ -56,7 +58,11 @@ read_stockholm_msa <- function(stockholm) {
             )
         )
     out <- attributes(seqs)
-    out[["alignment"]] <- Biostrings::RNAMultipleAlignment(unlist(seqs))
+    if (isTRUE(dna)) {
+        out[["alignment"]] <- Biostrings::DNAMultipleAlignment(unlist(seqs))
+    } else {
+        out[["alignment"]] <- Biostrings::RNAMultipleAlignment(unlist(seqs))
+    }
     out
 }
 
@@ -78,10 +84,73 @@ read_stockholm_msa <- function(stockholm) {
 #'   (global with respect to the model, local with respect to the sequence).
 #'   When \code{TRUE}, the search is faster, but will fail to find matches with
 #'   only partially homologous structure.
+#' @param Z (\code{numeric} scalar) Effective search space size in Mb for the
+#'   purposes of calculating E-values.
+#' @param output (\code{character} filename) File to send the human-readable
+#'   output to.
 #' @param alignment (filename) A file to save the aligned hits to.  If given,
 #'   the alignment is saved in Stockholm format with annotations for secondary
 #'   structure, posterior probablility, etc.
-#' @param cpu (integer of length 1) The number of threads to use in the search.
+#' @param acc (\code{logical} scalar) Use accessions instead of names in the
+#'   main output.
+#' @param noali (\code{logical} scalar) Omit the alignment section from the main
+#'   output.
+#' @param notextw (\code{logical} scalar) Unlimit the length of each line in the
+#'  main output.
+#' @param textw (\code{numeric} scalar) Set the main output’s line length limit
+#'   in characters per line. The default is 120.
+#' @param verbose (\code{logical} scalar)
+#' @param E (\code{numeric} scalar) Maximum E-value for reporting in per-target
+#'   output. Default: 10.0
+#' @param T (\code{numeric} scalar) Maximum bit score for reporting in
+#'   per-target reporting. Default: NULL
+#' @param incE (\code{numeric} scalar) Maximum E-value for hit inclusion.
+#'   Default: 0.01
+#' @param incT (\code{numeric} scalar) Maximum bit score for hit inclusion.
+#'   Default: NULL
+#' @param cut_ga (\code{logical} scalar) Use the GA (gathering) bit scores
+#'   defined in the CM to set hit reporting and inclusion thresholds.
+#' @param cut_nc (\code{logical} scalar) Use the NC (noise cutoff) bit score
+#'   thresholds defined in the CM to set hit reporting and inclusionthresholds.
+#' @param cut_tc (\code{logical} scalar) Use the TC (trusted cutoff) bit score
+#'   thresholds defined in the CM to set hit reporting and inclusionthresholds.
+#' @param filter_strategy (\code{character} string) Filtering strategy for the
+#'   acceleration pipeline. Options, from slowest (most sensitive) to fastest
+#'   (least sensitive) are "max", "nohmm", "mid", "default", "rfam", "hmmonly".
+#' @param FZ (\code{numeric} scalar) Effective database size in Mb for
+#'   determining filter thresholds.
+#' @param Fmid (\code{numeric} scalar) HMM filter thresholds for the "mid"
+#'   filtering strategy.  Default 0.02
+#' @param notrunc (\code{logical} scalar) Turn off truncated hit detection.
+#' @param anytrunc (\code{logical} scalar) Allow truncated hits to begin and end
+#'   at any position in a target sequence.
+#' @param nonull3 (\code{logical} scalar) Turn off the null3 CM score
+#'   corrections for biased composition.
+#' @param mxsize (\code{numeric} scalar) Maximum allowable CM DP matrix size in
+#'   megabytes. Default: 128
+#' @param smxsize (\code{numeric} scalar) Maximum allowable CM search DP matrix
+#'   size in megabytes. Default: 128
+#' @param cyk (\code{logical} scalar) Use the CYK algorithm, not Inside, to
+#'   determine the final score of all hits.
+#' @param acyk (\code{logical} scalar) Use the CYK algorithm to align hits.
+#' @param wcx (\code{numeric} scalar) Expected maximum length of a hit, as a
+#'   multiplier of consensus length of the model.
+#' @param toponly (\code{logical} scalar) Only search the top (Watson) strand of
+#'   target sequences.
+#' @param bottomonly (\code{logical} scalar) Only search the bottom (Crick)
+#'   strand of target sequences.
+#' @param tformat (\code{logical} scalar) Format of the sequences in "seq".
+#'   Options are "fasta", "embl", "genbank", "ddbj", "stockholm", "pfam", "a2m",
+#'   "afa", "clustal", and "phylip". Default: autodetect
+#' @param cpu (\code{integer} scalar) The number of threads to use in the
+#'   search.
+#' @param stall (\code{logical} scalar) For debugging the MPI master/worker
+#'   version: pause after start, to enable the developer toattach debuggers to
+#'   the running master and worker(s) processes
+#' @param mpi (\code{logical} scalar) Run  in  MPI  master/worker  mode, using
+#'   mpirun.
+#' @param quiet (\code{logical} scalar) Suppress standard output of `cmsearch`,
+#'   which can be long.
 #'
 #' @return a \code{\link[tibble]{tibble}} with columns:
 #'     \itemize{
@@ -118,25 +187,89 @@ read_stockholm_msa <- function(stockholm) {
 #'     # also works if the fasta file has already been loaded
 #'     samp <- Biostrings::readDNAStringSet(sampfasta)
 #'     cmsearch(cm = cm, seq = samp, cpu = 1)
-cmsearch <- function(cm, seq, glocal = TRUE, alignment = NULL, cpu = NULL) {
+cmsearch <- function(
+    cm,
+    seq,
+    glocal = TRUE,
+    Z = NULL,
+    output = NULL,
+    alignment = NULL,
+    acc = FALSE,
+    noali = TRUE,
+    notextw = FALSE,
+    textw = NULL,
+    verbose = FALSE,
+    E = NULL,
+    T = NULL,
+    incE = NULL,
+    incT = NULL,
+    cut_ga = FALSE,
+    cut_nc = FALSE,
+    cut_tc = FALSE,
+    filter_strategy = NULL,
+    FZ = NULL,
+    Fmid = NULL,
+    notrunc = FALSE,
+    anytrunc = FALSE,
+    nonull3 = FALSE,
+    mxsize = NULL,
+    smxsize = NULL,
+    cyk = FALSE,
+    acyk = FALSE,
+    wcx = NULL,
+    toponly = TRUE,
+    bottomonly = FALSE,
+    tformat = NULL,
+    cpu = NULL,
+    stall = FALSE,
+    mpi = FALSE,
+    quiet = TRUE) {
     assertthat::assert_that(assertthat::is.string(cm),
                             file.exists(cm),
                             assertthat::is.flag(glocal))
     tablefile <- tempfile("cmsearch", fileext = ".dat")
     on.exit(unlink(tablefile))
-    args <- c("--tblout", tablefile, "--toponly", "--noali")
-    if (isTRUE(glocal)) args <- c(args, "-g")
-    if (!is.null(cpu)) {
-        assertthat::assert_that(assertthat::is.count(cpu))
-        args <- c(args, "--cpu", cpu)
-    }
-    if (!is.null(alignment)) {
-        assertthat::assert_that(assertthat::is.string(alignment))
-        d <- dirname(alignment)
-        if (nchar(d) > 0 && !dir.exists(d)) dir.create(d, recursive = TRUE)
-        args <- c(args, "-A", alignment)
-    }
-    args <- c(args, cm)
+    args <- c(
+        flag_opt(glocal, "g"),
+        nonneg_float_opt(Z),
+        string_opt(output, "o"),
+        string_opt(alignment, "A"),
+        string_opt(tablefile, "tblout"),
+        flag_opt(acc),
+        flag_opt(noali),
+        flag_opt(notextw),
+        count_opt(textw),
+        flag_opt(verbose),
+        nonneg_float_opt(E),
+        nonneg_float_opt(T),
+        nonneg_float_opt(incE),
+        nonneg_float_opt(incT),
+        flag_opt(cut_ga),
+        flag_opt(cut_nc),
+        flag_opt(cut_tc),
+        multiflag_opt(filter_strategy,
+                      c("max", "nohmm", "mid", "default", "rfam", "hmmonly")),
+        nonneg_float_opt(FZ),
+        fraction_opt(Fmid),
+        flag_opt(notrunc),
+        flag_opt(anytrunc),
+        flag_opt(nonull3),
+        nonneg_float_opt(mxsize),
+        nonneg_float_opt(smxsize),
+        flag_opt(cyk),
+        flag_opt(acyk),
+        nonneg_float_opt(wcx),
+        flag_opt(toponly),
+        flag_opt(bottomonly),
+        string_opt(tformat, choices = c("fasta", "embl", "genbank", "ddbj",
+                                            "stockholm", "pfam", "a2m", "afa",
+                                            "clustal", "phylip")),
+        count_opt(cpu),
+        flag_opt(stall),
+        flag_opt(mpi),
+        cm
+    )
+
     seqfile <- NULL
     if (assertthat::is.string(seq) && file.exists(seq)) {
         seqfile <- seq
@@ -175,6 +308,7 @@ cmsearch <- function(cm, seq, glocal = TRUE, alignment = NULL, cpu = NULL) {
         col_types = "ccccciiiicciddddcc", comment = "#")
 }
 
+
 #' Align sequences to a covariance model
 #'
 #' This function calls \code{cmalign} from
@@ -184,8 +318,8 @@ cmsearch <- function(cm, seq, glocal = TRUE, alignment = NULL, cpu = NULL) {
 #' One of the easiest places to obtain CMs is
 #' \href{https://rfam.xfam.org/}{Rfam}.
 #'
-#' @param cmfile (\code{character} scalar) path to a covariance model file
-#' @param seq (\code{character} scalar, names \code{character} vector,
+#' @param cmfile (\code{character} filename) path to a covariance model file
+#' @param seq (\code{character} filename, \code{character} vector,
 #'        \code{\link[Biostrings]{XStringSet}}, or
 #'        \code{\link[ShortRead]{ShortRead}}) sequences to align to the
 #'        covariance model. This may be given as a character path to a fasta
@@ -199,12 +333,62 @@ cmsearch <- function(cm, seq, glocal = TRUE, alignment = NULL, cpu = NULL) {
 #' @param global (\code{logical} scalar) If \code{TRUE}, align in global mode.
 #'        See \href{http://eddylab.org/infernal/}{Infernal} documentation for
 #'        more information.
-#' @param cpu (\code{integer} scalar) The number of cpus to use.
-#' @param mxsize (\code{double} scalar) The maximum DP matrix size, in Mb.
+#' @param algorithm (\code{character} string) Alignment algorithm. Options are
+#'        "optacc" and "cyk". Default: "optacc"
+#' @param sample (\code{logical} scalar) Sample an alignment from the posterior
+#'        distribution of alignments.
+#' @param seed (\code{integer} scalar) Random seed for sampling an alignment.
+#' @param notrunc (\code{logical} scalar) Turn off truncated alignment
+#'        algorithms.
+#' @param sub (\code{logical} scalar) urn on the sub model construction and
+#'        alignment procedure
+#' @param hbanded (\code{logical} scalar) Accelerate alignment by pruning away
+#'        regions of the CMDP matrix that are deemed negligible by an HMM.
+#'        Default: TRUE
+#' @param tau (\code{numeric} scalar) Tail loss probability used during HMM band
+#'        calculation. Default: 1E-7
+#' @param mxsize (\code{numeric} scalar) The maximum DP matrix size, in Mb.
 #'        Maximum potential memory usage is approximately cpu*mxsize, although
 #'        this is usually not realized.
 #'        See \href{http://eddylab.org/infernal/}{Infernal} documentation for
 #'        more information.
+#' @param fixedtau (\code{logical} scalar) Turn off HMM band tightening.
+#' @param maxtau (\code{numeric} scalar) Maximum allowed value for tau
+#'        during band tightening. Default: 0.05
+#' @param small (\code{logical} scalar) Use the divide and conquer CYK alignment
+#'        algorithm, greatly reducing memory consumption.
+#' @param sfile (\code{character} filename) Dump per-sequence alignment score
+#'        and timing information to file.
+#' @param tfile (\code{character} filename) Dump tabular sequence tracebacks for
+#'        each individual sequence to a file.
+#' @param ifile (\code{character} filename) Dump per-sequence insert information
+#'        to file.
+#' @param elfile (\code{character} filename) Dump per-sequence EL state (local
+#'        end) insert information to file.
+#' @param mapali (\code{character} filename) Read the alignment from the file used
+#'        to build the model aligns it as a single object to the CM, along with
+#'        sequences in \code{"seq"}.
+#' @param mapstr (\code{logical} scalar) Must be used in combination with
+#'        \code{mapali}. Propogate structural information for any pseudoknots
+#'        that exist in \code{mapali} to the output alignment.
+#' @param dnaout (\code{logical} scalar) Output the alignments as DNA sequence
+#'        alignments, instead of RNA ones.
+#' @param noprob (\code{logical} scalar) Do not annotate the output alignment
+#'        with posterior probabilities.
+#' @param matchonly (\code{logical} scalar) Only include match columns in the
+#'        output alignment, do not include any insertions relativeto the
+#'        consensus model.
+#' @param ileaved (\code{logical} scalar) Output the alignment in interleaved
+#'        Stockholm format of a fixed width that may be more convenient for
+#'        examination.
+#' @param regress (\code{character} filename) Save an additional copy of the
+#'        output alignment with no author information to file.
+#' @param verbose (\code{logical} scalar) Output additional information in the
+#'        tabular scores output.
+#' @param cpu (\code{integer} scalar) The number of cpus to use.
+#' @param mpi (\code{logical} scalar) Run as an MPI parallel program.
+#' @param extra (\code{character} vector) Additional advanced options to pass
+#'        to `cmalign`.
 #' @param glocal (\code{logical} scalar) (Deprecated) see "global".
 #'
 #' @return the aligned sequences, as returned by
@@ -219,11 +403,39 @@ cmsearch <- function(cm, seq, glocal = TRUE, alignment = NULL, cpu = NULL) {
 #'     # also works if the fasta file has already been loaded
 #'     unaln <- Biostrings::readRNAStringSet(unaln)
 #'     cmalign(cm, unaln, cpu = 1)
-cmalign <- function(cmfile, seq, global = TRUE, cpu = NULL, mxsize = NULL,
-                    glocal = global) {
-    assertthat::assert_that(assertthat::is.readable(cmfile),
-                            assertthat::is.flag(glocal))
-    args <- "cmalign"
+cmalign <- function(
+    cmfile,
+    seq,
+    global = TRUE,
+    algorithm = NULL,
+    sample = FALSE,
+    seed = NULL,
+    notrunc = FALSE,
+    sub = FALSE,
+    hbanded = TRUE,
+    tau = NULL,
+    mxsize = NULL,
+    fixedtau = FALSE,
+    maxtau = NULL,
+    small = FALSE,
+    sfile = NULL,
+    tfile = NULL,
+    ifile = NULL,
+    elfile = NULL,
+    mapali = NULL,
+    mapstr = FALSE,
+    dnaout = FALSE,
+    noprob = FALSE,
+    matchonly = FALSE,
+    ileaved = FALSE,
+    regress = NULL,
+    verbose = FALSE,
+    cpu = NULL,
+    mpi = FALSE,
+    extra = NULL,
+    glocal = global
+) {
+    assertthat::assert_that(assertthat::is.readable(cmfile))
     if (!missing(glocal))  {
         warning("Use of 'glocal' in cmalign is deprecated. ",
                 "Please use 'global' instead.")
@@ -233,17 +445,37 @@ cmalign <- function(cmfile, seq, global = TRUE, cpu = NULL, mxsize = NULL,
         }
         global <- glocal
     }
-    if (isTRUE(global)) args <- c(args, "-g")
-    if (!is.null(cpu)) {
-        assertthat::assert_that(assertthat::is.count(cpu))
-        args <- c(args, "--cpu", cpu)
-    }
-    if (!is.null(mxsize)) {
-        assertthat::assert_that(
-            assertthat::is.number(mxsize),
-            mxsize > 0)
-        args <- c(args, "--mxsize", mxsize)
-    }
+
+    args <- c(
+        "cmalign",
+        flag_opt(global, "g"),
+        multiflag_opt(algorithm, c("optacc", "cyk")),
+        flag_opt(sample),
+        count_opt(seed),
+        flag_opt(notrunc),
+        flag_opt(sub),
+        flag_opt(hbanded, "nonbanded", invert = TRUE),
+        fraction_opt(tau),
+        count_opt(mxsize),
+        flag_opt(fixedtau),
+        fraction_opt(maxtau),
+        flag_opt(small),
+        string_opt(sfile),
+        string_opt(tfile),
+        string_opt(ifile),
+        string_opt(elfile),
+        infile_opt(mapali, ),
+        flag_opt(mapstr),
+        flag_opt(dnaout),
+        flag_opt(noprob),
+        flag_opt(matchonly),
+        flag_opt(ileaved),
+        string_opt(regress),
+        flag_opt(verbose),
+        count_opt(cpu),
+        flag_opt(mpi),
+        extra
+    )
 
     seqfile <- NULL
     if (assertthat::is.string(seq) && file.exists(seq)) {
@@ -272,9 +504,9 @@ cmalign <- function(cmfile, seq, global = TRUE, cpu = NULL, mxsize = NULL,
         }
     }
     args <- c(args, cmfile, seqfile)
-    args <- paste(args, collapse = " ")
+    args <- paste(shQuote(args), collapse = " ")
     alnpipe <- pipe(args)
-    read_stockholm_msa(alnpipe)
+    read_stockholm_msa(alnpipe, dna = dnaout)
 }
 
 #' Construct covariance model(s) from structually annotated alignment(s)